refactor(cli): extract the wifi-bridge network harness out of main.rs
main.rs held ~460 lines of 802.11 / DHCP / ARP / UDP frame builders, parsers,
and the virtual-AP responder for the wifi-bridge command — network simulation
logic in the CLI entry point. Move it to a focused crates/cli/src/wifi_frames.rs
module; main.rs keeps command dispatch (6318 -> 5859 lines).

Pure code move (CLI-only; core untouched). The esp32c3 virtual_wifi peripheral
keeps its own station-side frame helpers — they share names but have diverged
signatures, so they are deliberately NOT merged here.

Verified: labwired-cli build + clippy clean, workspace lib green.
